requirements.txt
Django
psycopg2-binary
djangorestframework
python-decouple
requests